rename: 채팅 목록 페이지 위치 이동 (#201)
